requirements.txt
seaborn~=0.12.2
pandas~=1.5.3
matplotlib~=3.7.2
numpy~=1.25.2
bokeh~=3.2.1
scipy~=1.11.1
pillow~=10.0.0
scikit-learn~=1.3.0
natsort~=7.1.1
torch~=2.0.1
tqdm~=4.65.0
torchvision~=0.15.2